Task7/requirements.txt
streamlit
textblob
google-generativeai
python-dotenv
Pillow
